services/data_extractor/requirements.txt
# Data Extractor Service Requirements

# File extraction libraries
PyPDF2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.4

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-mock==3.14.0
